backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
supabase==2.3.4
apscheduler==3.10.4
scikit-learn==1.4.0
shap==0.44.1
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
sib-api-v3-sdk==7.6.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
joblib==1.3.2
email-validator==2.3.0